feat: cartesian chart - test cases updated
