+++ projects/observability/src/pages/apis/api-detail/overview/api-overview.dashboard.ts
@@ -297,7 +297,8 @@ export const apiOverviewDashboard: DashboardDefaultConfiguration = {
               }
             ],
             'selection-handler': {
-              type: 'cartesian-chart-selection-handler'
+              type: 'cartesian-explorer-selection-handler',
+              'show-context-menu': true
             }
           },
           {
@@ -394,7 +395,8 @@ export const apiOverviewDashboard: DashboardDefaultConfiguration = {
               }
             ],
             'selection-handler': {
-              type: 'cartesian-chart-selection-handler'
+              type: 'cartesian-explorer-selection-handler',
+              'show-context-menu': true
             }
           },
           {
@@ -491,7 +493,8 @@ export const apiOverviewDashboard: DashboardDefaultConfiguration = {
               }
             ],
             'selection-handler': {
-              type: 'cartesian-chart-selection-handler'
+              type: 'cartesian-explorer-selection-handler',
+              'show-context-menu': true
             }
           }
         ]
