+++ projects/observability/src/pages/apis/backend-detail/metrics/backend-metrics.dashboard.ts
@@ -256,7 +256,8 @@ export const backendMetricsDashboard: DashboardDefaultConfiguration = {
               }
             ],
             'selection-handler': {
-              type: 'cartesian-chart-selection-handler'
+              type: 'cartesian-explorer-selection-handler',
+              'show-context-menu': true
             }
           },
           {
@@ -319,7 +320,8 @@ export const backendMetricsDashboard: DashboardDefaultConfiguration = {
                   }
                 ],
                 'selection-handler': {
-                  type: 'cartesian-chart-selection-handler'
+                  type: 'cartesian-explorer-selection-handler',
+                  'show-context-menu': true
                 }
               },
               {
@@ -345,7 +347,8 @@ export const backendMetricsDashboard: DashboardDefaultConfiguration = {
                   }
                 ],
                 'selection-handler': {
-                  type: 'cartesian-chart-selection-handler'
+                  type: 'cartesian-explorer-selection-handler',
+                  'show-context-menu': true
                 }
               }
             ]
@@ -377,7 +380,8 @@ export const backendMetricsDashboard: DashboardDefaultConfiguration = {
           }
         ],
         'selection-handler': {
-          type: 'cartesian-chart-selection-handler'
+          type: 'cartesian-explorer-selection-handler',
+          'show-context-menu': true
         }
       }
     ]
