+++ projects/observability/src/pages/apis/backend-detail/overview/backend-overview.component.ts
@@ -289,7 +289,8 @@ export class BackendOverviewComponent {
               }
             ],
             'selection-handler': {
-              type: 'cartesian-chart-selection-handler'
+              type: 'cartesian-explorer-selection-handler',
+              'show-context-menu': true
             }
           },
           {
@@ -370,7 +371,8 @@ export class BackendOverviewComponent {
               }
             ],
             'selection-handler': {
-              type: 'cartesian-chart-selection-handler'
+              type: 'cartesian-explorer-selection-handler',
+              'show-context-menu': true
             }
           },
           {
@@ -451,7 +453,8 @@ export class BackendOverviewComponent {
               }
             ],
             'selection-handler': {
-              type: 'cartesian-chart-selection-handler'
+              type: 'cartesian-explorer-selection-handler',
+              'show-context-menu': true
             }
           }
         ]
