+++ projects/observability/src/pages/apis/service-detail/overview/service-overview.dashboard.ts
@@ -341,7 +341,8 @@ export const serviceOverviewDashboard: DashboardDefaultConfiguration = {
                   }
                 ],
                 'selection-handler': {
-                  type: 'cartesian-chart-selection-handler'
+                  type: 'cartesian-explorer-selection-handler',
+                  'show-context-menu': true
                 }
               },
               {
@@ -438,7 +439,8 @@ export const serviceOverviewDashboard: DashboardDefaultConfiguration = {
                   }
                 ],
                 'selection-handler': {
-                  type: 'cartesian-chart-selection-handler'
+                  type: 'cartesian-explorer-selection-handler',
+                  'show-context-menu': true
                 }
               },
               {
@@ -535,7 +537,8 @@ export const serviceOverviewDashboard: DashboardDefaultConfiguration = {
                   }
                 ],
                 'selection-handler': {
-                  type: 'cartesian-chart-selection-handler'
+                  type: 'cartesian-explorer-selection-handler',
+                  'show-context-menu': true
                 }
               }
             ]
