+++ projects/observability/src/pages/explorer/explorer-dashboard-builder.test.ts
@@ -43,7 +43,8 @@ describe('Explorer dashboard builder', () => {
             'series-from-data': true,
             'legend-position': LegendPosition.Bottom,
             'selection-handler': {
-              type: 'cartesian-explorer-selection-handler'
+              type: 'cartesian-explorer-selection-handler',
+              'show-context-menu': false
             }
           },
           onReady: expect.any(Function)
