+++ projects/observability/src/pages/explorer/explorer-dashboard-builder.ts
@@ -64,7 +64,8 @@ export class ExplorerDashboardBuilder {
         'series-from-data': true,
         'legend-position': LegendPosition.Bottom,
         'selection-handler': {
-          type: 'cartesian-explorer-selection-handler'
+          type: 'cartesian-explorer-selection-handler',
+          'show-context-menu': false
         }
       },
       onReady: dashboard => {
