+++ projects/observability/src/shared/dashboard/widgets/charts/cartesian-widget/interactions/cartesian-explorer-selection-handler.model.ts
@@ -1,9 +1,16 @@
-import { PopoverRef } from '@hypertrace/components';
-import { Model } from '@hypertrace/hyperdash';
+import {
+  PopoverBackdrop,
+  PopoverFixedPositionLocation,
+  PopoverPositionType,
+  PopoverRef,
+  PopoverService
+} from '@hypertrace/components';
+import { BOOLEAN_PROPERTY, Model, ModelProperty } from '@hypertrace/hyperdash';
 import { ModelInject } from '@hypertrace/hyperdash-angular';
 import { Observable, of } from 'rxjs';
 import { CartesianSelectedData } from '../../../../../../public-api';
 import { InteractionHandler } from '../../../../interaction/interaction-handler';
+import { CartesianExplorerContextMenuComponent } from './cartesian-explorer-context-menu/cartesian-explorer-context-menu.component';
 
 import { CartesainExplorerNavigationService } from './cartesian-explorer-navigation.service';
 
@@ -14,14 +21,43 @@ export class CartesianExplorerSelectionHandlerModel<TData> implements Interactio
   @ModelInject(CartesainExplorerNavigationService)
   private readonly cartesainExplorerNavigationService!: CartesainExplorerNavigationService;
 
+  @ModelInject(PopoverService)
+  private readonly popoverService!: PopoverService;
+
   public popover?: PopoverRef;
 
+  @ModelProperty({
+    key: 'show-context-menu',
+    displayName: 'Show Context Menu',
+    type: BOOLEAN_PROPERTY.type
+  })
+  public isContextMenuVisible: boolean = true;
+
   public execute(selectionData: CartesianSelectedData<TData>): Observable<void> {
-    this.cartesainExplorerNavigationService.navigateToExplorer(
-      selectionData.timeRange.startTime,
-      selectionData.timeRange.endTime
-    );
+    if (this.isContextMenuVisible) {
+      this.showContextMenu(selectionData);
+      this.popover?.closeOnBackdropClick();
+      this.popover?.closeOnPopoverContentClick();
+    } else {
+      this.cartesainExplorerNavigationService.navigateToExplorer(
+        selectionData.timeRange.startTime,
+        selectionData.timeRange.endTime
+      );
+    }
 
     return of();
+  }
+
+  public showContextMenu(selectionData: CartesianSelectedData<TData>): void {
+    this.popover = this.popoverService.drawPopover({
+      componentOrTemplate: CartesianExplorerContextMenuComponent,
+      data: selectionData,
+      position: {
+        type: PopoverPositionType.Fixed,
+        location: PopoverFixedPositionLocation.Custom,
+        customLocation: selectionData.location
+      },
+      backdrop: PopoverBackdrop.Transparent
+    });
   }
 }
